files: move machine_token to a separate module
Create specific module for the machine token file
