+++ uaclient/files/__init__.py
@@ -1,5 +1,6 @@
 from uaclient.files.data_types import DataObjectFile, DataObjectFileFormat
-from uaclient.files.files import MachineTokenFile, UAFile
+from uaclient.files.files import UAFile
+from uaclient.files.machine_token import MachineTokenFile
 
 __all__ = [
     "DataObjectFile",
